feat: extracts filter logic to useFilters hook
