+++ packages/cloud-cognitive/src/components/Datagrid/Datagrid/addons/Filtering/hooks/index.js
@@ -8,3 +8,4 @@
 export { default as useInitialStateFromFilters } from './useInitialStateFromFilters';
 export { default as useSubscribeToEventEmitter } from './useSubscribeToEventEmitter';
 export { default as useShouldDisableButtons } from './useShouldDisableButtons';
+export { default as useFilters } from './useFilters';
